requirements.txt
anyio==4.4.0
asgiref==3.8.1
blinker==1.8.2
chardet==5.2.0
click==8.1.7
Flask==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
pillow==10.4.0
python-dotenv==1.0.1
PyYAML==6.0.1
reportlab==4.2.2
sniffio==1.3.1
uvicorn==0.30.1
watchfiles==0.22.0
Werkzeug==3.0.3
